Augment Function Hyperbolic

Takes a numeric vector(s) or date and will return a tibble of one of the following:

"sin"

"cos"

"tan"

"sincos"

c("sin","cos","tan", "sincos")

hai_hyperbolic_augment(
  .data,
  .value,
  .names = "auto",
  .scale_type = c("sin", "cos", "tan", "sincos")
)

Arguments

.data
The data being passed that will be augmented by the function.

.value
This is passed rlang::enquo() to capture the vectors you want to augment.

.names
The default is "auto"

.scale_type
A character of one of the following: "sin","cos","tan", "sincos" All can be passed by setting the param equal to c("sin","cos","tan","sincos")

Value

A augmented tibble

Details

Takes a numeric vector or date and will return a vector of one of the following:

"sin"

"cos"

"tan"

"sincos"

c("sin","cos","tan", "sincos")

This function is intended to be used on its own in order to add columns to a tibble.

See also

Author

Steven P. Sanderson II, MPH

Examples

suppressPackageStartupMessages(library(dplyr))

len_out    = 10
by_unit    = "month"
start_date = as.Date("2021-01-01")

data_tbl <- tibble(
  date_col = seq.Date(from = start_date, length.out = len_out, by = by_unit),
  a    = rnorm(len_out),
  b    = runif(len_out)
)

hai_hyperbolic_augment(data_tbl, b, .scale_type = "sin")
#> # A tibble: 10 x 4
#>    date_col        a     b hyperbolic_b_sin
#>    <date>      <dbl> <dbl>            <dbl>
#>  1 2021-01-01 -1.41  0.578            0.547
#>  2 2021-02-01 -0.506 0.242            0.239
#>  3 2021-03-01 -0.270 0.114            0.114
#>  4 2021-04-01 -1.09  0.298            0.294
#>  5 2021-05-01  0.362 0.484            0.465
#>  6 2021-06-01 -0.336 0.567            0.537
#>  7 2021-07-01  1.36  0.754            0.685
#>  8 2021-08-01 -0.712 0.545            0.519
#>  9 2021-09-01  0.662 0.760            0.689
#> 10 2021-10-01  0.291 0.213            0.211
hai_hyperbolic_augment(data_tbl, b, .scale_type = "tan")
#> # A tibble: 10 x 4
#>    date_col        a     b hyperbolic_b_tan
#>    <date>      <dbl> <dbl>            <dbl>
#>  1 2021-01-01 -1.41  0.578            0.653
#>  2 2021-02-01 -0.506 0.242            0.246
#>  3 2021-03-01 -0.270 0.114            0.115
#>  4 2021-04-01 -1.09  0.298            0.308
#>  5 2021-05-01  0.362 0.484            0.526
#>  6 2021-06-01 -0.336 0.567            0.637
#>  7 2021-07-01  1.36  0.754            0.939
#>  8 2021-08-01 -0.712 0.545            0.607
#>  9 2021-09-01  0.662 0.760            0.950
#> 10 2021-10-01  0.291 0.213            0.216