Get the Juiced Data

This is a simple function that will get the juiced data from a recipe.

Usage

get_juiced_data(.recipe_object)

Arguments

.recipe_object

The recipe object you want to pass.

Value

A tibble of the prepped and juiced data from the given recipe

Details

Instead of typing out something like: recipe_object %>% prep() %>% juice() %>% glimpse()

See also

Author

Steven P. Sanderson II, MPH

Examples

suppressPackageStartupMessages(library(timetk))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(purrr))
suppressPackageStartupMessages(library(healthyR.data))
suppressPackageStartupMessages(library(rsample))
suppressPackageStartupMessages(library(recipes))

data_tbl <- healthyR_data %>%
    select(visit_end_date_time) %>%
    summarise_by_time(
        .date_var = visit_end_date_time,
        .by       = "month",
        value     = n()
    ) %>%
    set_names("date_col","value") %>%
    filter_by_time(
        .date_var = date_col,
        .start_date = "2013",
        .end_date = "2020"
    )

splits <- initial_split(data = data_tbl, prop = 0.8)

rec_obj <- recipe(value ~., training(splits))

get_juiced_data(rec_obj)
#> # A tibble: 76 × 2
#>    date_col            value
#>    <dttm>              <int>
#>  1 2013-02-01 00:00:00  1719
#>  2 2019-05-01 00:00:00  1486
#>  3 2017-07-01 00:00:00  1535
#>  4 2013-01-01 00:00:00  2082
#>  5 2015-02-01 00:00:00  1452
#>  6 2013-12-01 00:00:00  1670
#>  7 2015-11-01 00:00:00  1447
#>  8 2018-02-01 00:00:00  1514
#>  9 2017-09-01 00:00:00  1512
#> 10 2014-06-01 00:00:00  1753
#> # … with 66 more rows